Statically link OpenSSL
I've been back and forth with this before but compiling OpenSSL means I can support more targets and anyone that uses komac on Linux or in automation doesn't have to install libssl beforehand. It also means that the OpenSSL version komac uses will be the latest as of when komac was built.