Request less data when retrieving package versions
Addresses https://github.com/Jackett/Jackett/issues/13948#issuecomment-2302906882.

Before this, Komac used GraphQL recursion one level deep to retrieve a package's directory content so that there are less structs to declare. However, at the base directory level, only the folder name is required and at one level deep, only the file type is required.

Now, the query is more precise and requests only the data that is needed albeit requiring a few more structs to declare this.

This was necessary because GitHub was timing out getting the versions of packages with a lot of versions, such as `Jackett.Jackett`. Retrieving versions is now also noticeably quicker.