Get recursive tree for retrieving package versions
This uses GitHub's rest API instead of GraphQL as there isn't a good way to do recursion in GraphQL that take longer and time out.

This resolves the `Something went wrong while executing your query. This may be the result of a timeout, or it could be a GitHub bug` error when retrieving packages with a lot of versions.