Get installer locale from NSIS installers
This required an update to `zerocopy` 0.8.0 (currently alpha) as this adds support for dynamically-sized-types (DSTs) such as `[u8]` which is needed for the NSIS language table that contains a variable-length array of language string pointers.

`zerocopy` 0.8.0 also adds `TryFromBytes` so the NSIS FirstHeader can be fully parsed using zerocopy and will throw an error if any of the magic bytes aren't correct.

This commit also improves NSIS version detection by checking for NSIS 2 and NSIS 3 codes at the start of each string in the string block. Before, it was operating on an incorrect assumption that the branding text is always `Nullsoft Scriptable Install System v` but this is not always the case.

This commit also adds a check for if the install directory pointer is 0 as this may mean that the install directory has not been defined or that the header struct has been reordered to hinder decompilation.