Fallback to getting NSIS version from branding text
This was removed in https://github.com/russellbanks/Komac/commit/fc89acd0eb9e1f449a95bee36abfc22fb89fdf66. This has been re-added here with a more strict check as it can act as a fallback for when the installer does not have a manifest while still getting a more accurate version than checking the codes of the string block.