Deduplicate installers
This bug would have been introduced when the use of a BTreeSet was replaced with a Vec so that installers could be reordered manually and more efficiently (necessary because you can't mutably modify objects inside a BTree - https://stackoverflow.com/a/58790477). This Vec was only being sorted though and duplicates weren't being removed so it wasn't fully mimicking the functionality of the previous BTreeSet.

Resolves #958