Add `--hash` flag to analyse command
This is the command I use the most for debugging but almost all of the time taken to run it is from hashing the file. This is mitigated in other commands because the file is being hashed while it's downloading.

This makes the default behaviour to not hash the file. Using `--hash` will hash the file.